refactor: move the wasmbinding away from the tokenfactory module
