+++ wasmbinding/testdata/token_reflect.wasm
