+++ wasmbinding/types/query.go
